+++ impexp-kml-collada-plugin/src/main/java/org/citydb/modules/kml/database/CityObjectGroup.java
@@ -98,7 +98,7 @@ public class CityObjectGroup extends KmlGenericObject{
 			String query = queries.getCityObjectGroupFootprint();
 			psQuery = connection.prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
 			for (int i = 1; i <= getParameterCount(query); i++)
-				psQuery.setLong(i, work.getId());
+				psQuery.setLong(i, (long)work.getId());
 
 			rs = psQuery.executeQuery();
 			if (!rs.isBeforeFirst()) {
