+++ impexp-kml-collada-plugin/src/main/java/org/citydb/modules/kml/database/Transportation.java
@@ -117,7 +117,7 @@ public class Transportation extends KmlGenericObject{
 					String query = queries.getTransportationQuery(currentLod, work.getDisplayForm());				
 					psQuery = connection.prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
 					for (int i = 1; i <= getParameterCount(query); i++)
-						psQuery.setLong(i, work.getId());
+						psQuery.setLong(i, (long)work.getId());
 
 					rs = psQuery.executeQuery();
 					if (rs.isBeforeFirst())
@@ -175,7 +175,7 @@ public class Transportation extends KmlGenericObject{
 							String query = queries.getExtrusionHeight();
 							psQuery2 = connection.prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
 							for (int i = 1; i <= getParameterCount(query); i++)
-								psQuery2.setLong(i, work.getId());
+								psQuery2.setLong(i, (long)work.getId());
 
 							rs2 = psQuery2.executeQuery();
 							rs2.next();
@@ -213,9 +213,9 @@ public class Transportation extends KmlGenericObject{
 							log.info("Object " + work.getGmlId() + " has more than " + GEOMETRY_AMOUNT_WARNING + " geometries. This may take a while to process...");
 
 						List<Point3d> anchorCandidates = getOrigins(); // setOrigins() called mainly for the side-effect
-						double zOffset = getZOffsetFromConfigOrDB(work.getId());
+						double zOffset = getZOffsetFromConfigOrDB((long)work.getId());
 						if (zOffset == Double.MAX_VALUE) {
-							zOffset = getZOffsetFromGEService(work.getId(), anchorCandidates);
+							zOffset = getZOffsetFromGEService((long)work.getId(), anchorCandidates);
 						}
 						setZOffset(zOffset);
 
@@ -266,11 +266,11 @@ public class Transportation extends KmlGenericObject{
 
 		List<PlacemarkType> placemarkList= new ArrayList<>();
 		
-		double zOffset = getZOffsetFromConfigOrDB(work.getId());
+		double zOffset = getZOffsetFromConfigOrDB((long)work.getId());
 		List<Point3d> lowestPointCandidates = getLowestPointsCoordinates(rs, (zOffset == Double.MAX_VALUE));
 		rs.beforeFirst(); // return cursor to beginning
 		if (zOffset == Double.MAX_VALUE) {
-			zOffset = getZOffsetFromGEService(work.getId(), lowestPointCandidates);
+			zOffset = getZOffsetFromGEService((long)work.getId(), lowestPointCandidates);
 		}
 
 		while (rs.next()) {
@@ -330,7 +330,7 @@ public class Transportation extends KmlGenericObject{
 			}
 			// replace default BalloonTemplateHandler with a brand new one, this costs resources!
 			if (getBalloonSettings().isIncludeDescription()) {
-				addBalloonContents(placemark, work.getId());
+				addBalloonContents(placemark, (long)work.getId());
 			}
 			placemarkList.add(placemark);
 		}
